packages/db/pyproject.toml
[project]
name = "db-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.13.3",
    "asyncpg>=0.31.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.1.1",
    "tqdm>=4.67.1",
    "unidecode>=1.4.0",
    "alembic (>=1.16.4,<2.0.0)",
    "anyio (>=4.12.1,<5)",
    "greenlet>=3.3.2",
]

[tool.uv]
package = true
